# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoroBeacon API",
    "description": "Monitoring and alerting for Stellar/Soroban contract events. All endpoints are mounted under /api/v1. The API is read-write for monitors, rules and channels; alerts are read-only.\n\nAuthentication: when API_TOKEN is configured, every endpoint below requires `Authorization: Bearer <token>` and answers 401 through the standard error envelope without one. The only exceptions are the probes /livez, /readyz and /health, which stay reachable so an orchestrator does not need the token. With API_TOKEN unset the API is open and none of this applies.",
    "version": "1.0.0",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
    "/monitors": {
      "post": {
        "description": "Required role: editor",
        "summary": "Create a monitor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MonitorInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Monitor" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      },
      "get": {
        "description": "Required role: viewer",
        "summary": "List monitors",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Case-insensitive substring match against the monitor name"
          },
          {
            "name": "enabled",
            "in": "query",
            "schema": { "type": "boolean" },
            "description": "Tri-state: omitted = all, true = enabled only, false = disabled only"
          },
          {
            "name": "sort",
            "in": "query",
            "schema": { "type": "string", "enum": ["name", "id", "created_at"], "default": "name" },
            "description": "Allowlisted order: name (default), id (newest first), created_at"
          },
          {
            "name": "network",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Only monitors watching this Stellar chain. Omitted means every chain this instance polls."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Opaque continuation cursor from a previous page"
          }
        ],
        "responses": {
          "200": {
            "description": "Monitors",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MonitorPage" }
              }
            }
          }
        }
      }
    },
    "/monitors/bulk": {
      "post": {
        "description": "Required role: editor",
        "summary": "Bulk enable or disable monitors",
        "description": "Sets enabled on an explicit list of monitor IDs in one statement. Unknown IDs are returned and do not prevent the known ones from applying. There is no all-monitors shorthand.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BulkMonitorInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Applied",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BulkMonitorResult" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/monitors/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ID" }
      ],
      "get": {
        "summary": "Fetch one monitor",
        "responses": {
          "200": {
            "description": "Monitor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Monitor" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "summary": "Update a monitor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MonitorInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Monitor" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "summary": "Delete a monitor",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/channels": {
      "post": {
        "summary": "Create a notification channel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChannelInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Channel" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      },
      "get": {
        "summary": "List channels",
        "parameters": [
          {
            "name": "enabled",
            "in": "query",
            "schema": { "type": "boolean" },
            "description": "When true, only enabled channels are returned"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Opaque continuation cursor from a previous page"
          }
        ],
        "responses": {
          "200": {
            "description": "Channels",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChannelPage" }
              }
            }
          }
        }
      }
    },
    "/channels/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ID" }
      ],
      "get": {
        "summary": "Fetch one channel",
        "responses": {
          "200": {
            "description": "Channel",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Channel" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "summary": "Update a channel",
        "description": "A partial update. Setting `enabled: true` on a channel that was off is the only way out of an auto-disable, and clears its failure counters in the same write so it does not immediately re-disable.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChannelInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Channel" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "summary": "Delete a channel",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/channels/{id}/test": {
      "parameters": [
        { "$ref": "#/components/parameters/ID" }
      ],
      "post": {
        "summary": "Send a test notification through a channel",
        "responses": {
          "200": { "description": "Test sent" },
          "404": { "$ref": "#/components/responses/Error" },
          "502": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/maintenance-windows": {
      "post": {
        "summary": "Create a maintenance window",
        "description": "Time-bounded silence. Alerts raised inside the window are still persisted and visible but are not delivered.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MaintenanceWindowInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MaintenanceWindow" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      },
      "get": {
        "summary": "List maintenance windows",
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "schema": { "type": "boolean" },
            "description": "Restrict to windows containing the current time"
          },
          {
            "name": "upcoming",
            "in": "query",
            "schema": { "type": "boolean" },
            "description": "Restrict to windows starting in the future"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 500, "default": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance windows",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MaintenanceWindowPage" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/maintenance-windows/{id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ID" }
      ],
      "get": {
        "summary": "Fetch one maintenance window",
        "responses": {
          "200": {
            "description": "Maintenance window",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MaintenanceWindow" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "summary": "Update a maintenance window",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MaintenanceWindowInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MaintenanceWindow" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "summary": "Delete a maintenance window",
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/inhibitions/{sourceID}/{targetID}": {
      "delete": {
        "summary": "Delete an inhibition rule",
        "parameters": [
          {
            "name": "sourceID",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "targetID",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Any of three credentials. (1) A static token from API_TOKEN (comma-separated to allow rotation): full reach over the default workspace. (2) A scoped token minted by POST /tokens — it looks like `sb_…`, and carries a scope list, an optional expiry and a revocation. Each route names the scopes it requires; a route that has no mapping denies a scoped token with 403 rather than allowing it. (3) A dashboard session cookie from /login, accepted in the browser's own same-origin requests. A token from WORKSPACE_TOKENS (`workspace=token` pairs) is accepted the same way as (1) but additionally scopes the request: it sees and changes only that workspace's monitors, channels, rules, alerts, saved searches and templates. `API_TOKEN`'s tokens belong to the `default` workspace, as do rows created before workspaces existed. Scoped tokens are workspace-bound too: the workspace comes from the row that stores the token's digest, not from the request. There is deliberately no X-Workspace header or query parameter — the workspace is a property of the credential, so a client cannot ask for another tenant's data."
      }
    },
...
  }
}